requirements.txt
asgiref==3.11.0
black==25.12.0
click==8.3.1
colorama==0.4.6
coverage==7.13.0
Django==6.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-yasg==1.21.11
flake8==7.3.0
inflection==0.5.1
isort==7.0.0
mccabe==0.7.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
pillow==12.0.0
platformdirs==4.5.1
psycopg2-binary==2.9.11
pycodestyle==2.14.0
pyflakes==3.4.0
PyJWT==2.10.1
pytokens==0.3.0
pytz==2025.2
PyYAML==6.0.3
sqlparse==0.5.4
tzdata==2025.3
uritemplate==4.2.0
